Moved CFG to own header and implementation (wip)
